Add Bacon::Context.setup and separate the setup phase from the run phase.
Instead of running the specs directly from the it and describe calls, I
added Bacon::Context.setup for setting up contexts and examples (and storing
them in instance variables). After that, bin/bacon takes care of running the
contexts by calling Bacon.contexts.each { |context| context.run }.

This means Counter[:specifications] is set _before_ the tests are actually
run (in the setup phase), so you know how many tests you're going to run. Also,
this could allow you to change the order in which the tests are run.

NOTE: This is a prototype, all original tests pass (except for that one in
test/spec_should) but this should definitely be rewritten.